PIDL Object ids need to be large enough to accomodate a 64-bit pointer.  Chnaged from int to size_t.

+++ src/Core/CCA/PIDL/ServerContext.h
@@ -80,7 +80,7 @@ DESCRIPTION
     //////////
     // The ID of this object from the object wharehouse.  This
     // id is unique within this process.
-    int d_objid;
+    size_t d_objid;
 
     //////////
     // A pointer to the object base class.
